+++ engine/telemetry/tracing/tracing_test.go
@@ -1,3 +1,0 @@
-package tracing
-
-// Tracing public tests removed (C8). Package slated for deletion.